resolve issues on PR
