+++ src/hooks/useAccount.ts
@@ -1,4 +1,3 @@
-import { useIsMountedRef } from './useIsMountedRef';
 import { reactive, toRefs, watchEffect } from 'vue';
 import { u8aToHex } from '@polkadot/util';
 import { keyring } from '@polkadot/ui-keyring';
@@ -17,7 +16,7 @@ interface UseAccounts {
  * The hooks should only be called from the `setup()` block of the vue component
  */
 export const useAccount = () => {
-  const mountedRef = useIsMountedRef();
+  // const mountedRef = useIsMountedRef();
 
   // set the initial value
   const state = reactive<UseAccounts>({
@@ -39,22 +38,20 @@ export const useAccount = () => {
     // fixme: this part is showing an error when fetching accounts
     const subscription = keyring.accounts.subject.subscribe((accounts) => {
       // only subscribe to the keyring if the component that originally called this hook is still mounted
-      if (mountedRef.value) {
-        // fixme: this is an unintuitive method to assign values. We need to find a scalable method
-        state.allAccounts = accounts ? Object.keys(accounts) : [];
-        state.allAccountNames = accounts
-          ? Object.values(accounts).map((obj) => obj.option.name)
-          : [];
-        state.defaultAccount =
-          state.allAccounts.length > 0 ? Object.keys(accounts)[0] : '';
-        state.defaultAccountName =
-          state.allAccounts.length > 0
-            ? Object.values(accounts)[0].option.name
-            : '';
-        state.hasAccounts = state.allAccounts.length !== 0;
-        state.isAccount = (address: string) =>
-          state.allAccounts.includes(address);
-      }
+      // fixme: this is an unintuitive method to assign values. We need to find a scalable method
+      state.allAccounts = accounts ? Object.keys(accounts) : [];
+      state.allAccountNames = accounts
+        ? Object.values(accounts).map((obj) => obj.option.name)
+        : [];
+      state.defaultAccount =
+        state.allAccounts.length > 0 ? Object.keys(accounts)[0] : '';
+      state.defaultAccountName =
+        state.allAccounts.length > 0
+          ? Object.values(accounts)[0].option.name
+          : '';
+      state.hasAccounts = state.allAccounts.length !== 0;
+      state.isAccount = (address: string) =>
+        state.allAccounts.includes(address);
     });
 
     // unsubscribe from the keyring if the parent component is unmounted
